Refactor lazy loading comments to allow for edge cases (#1647)
* Refactor lazy loading comments to allow for edge cases

Previously scenarios in which the user managed to "avoid" the
intersection observer (either scrolling too fast, or refreshing
from the bottom of the page) the comments would fail to load.

This change adds an additional check which happens on the very
first page scroll event (also fired when refreshed). If the check
finds the user is already below the intersection we would have
observed then we load the comments and dont init the observer.